+++ src/2-pages/Stats/WidgetNetWorth/WidgetHeader.tsx
@@ -5,24 +5,16 @@ import { Period, PeriodTitle } from '../shared/period'
 import { useTranslation } from 'react-i18next'
 import { formatMoney } from '6-shared/helpers/money'
 import { displayCurrency } from '5-entities/currency/displayCurrency'
-import { useStatSummary } from '../WidgetStatCards/model'
+import { useStatSummary } from '../shared/cashflow'
 import { Tooltip } from '6-shared/ui/Tooltip'
 import { SavingsTooltip } from './SavingsTooltip'
 import { getSavingsBackgroundColor } from "6-shared/ui/theme/colors";
 
+const percentThreshold = 0.05
+
 type WidgetHeaderProps = {
   period: Period
   onTogglePeriod: () => void
-}
-
-const PERCENT_THRESHOLD = 0.05
-
-function useFormatters(currency: string) {
-  return useMemo(() => ({
-    formatCurrency: (amount: number): string => formatMoney(amount, currency),
-    formatPercent: (value: number): string =>
-      Math.abs(value) < PERCENT_THRESHOLD ? '0.0' : value.toFixed(1)
-  }), [currency])
 }
 
 export function WidgetHeader({ period, onTogglePeriod }: WidgetHeaderProps) {
@@ -56,10 +48,10 @@ export function WidgetHeader({ period, onTogglePeriod }: WidgetHeaderProps) {
               stats={stats}
               period={period}
               formatCurrency={formatCurrency}
-              formatPercent={formatPercent}
             />
           }
           arrow
+          placement="right"
         >
           <Typography
             component="span"
@@ -88,4 +80,14 @@ export function WidgetHeader({ period, onTogglePeriod }: WidgetHeaderProps) {
       </Typography>
     </Box>
   )
+}
+
+function useFormatters(currency: string) {
+  return useMemo(() => ({
+    formatCurrency: (amount: number): string => formatMoney(amount, currency),
+    formatPercent: (value: number): string => {
+      if (Math.abs(value) < percentThreshold) return '0.0'
+      return value.toFixed(Math.abs(value) < 1 ? 2 : 1)
+    }
+  }), [currency])
 }
