--- src/2-pages/Stats/WidgetStatCards/index.tsx
@@ -1,1 +1,0 @@
-export { WidgetStatCards } from './WidgetStatCards'
